Move GCPTokenStore to separate file, out from config.py
